refactor (core): use common geometry in meshTile
Identical geometries are reused to build the tileMeshs.
The normal in the center of the geometry is aligned with the Z axis.
The bounding boxes is calculated by THREE.js.
Bounding boxes can use the Parent's worldMatrix (tile).
reduces obb calculation errors see #87 issues
The geometry normal is computed with correct geodesic normal.
fix obb/culling for pano tile